alpha

+++ src/bitvector.cpp
@@ -1,7 +1,7 @@
 #include <stdexcept>
 #include "bitvector.hpp"
 
-void BitVector::append (const unsigned int &value, const size_t &length) {
+void BitVector::append (const size_t &value, const size_t &length) {
 	if (length > 31 || value >> length != 0) {
     throw std::out_of_range("BitVector length argument");
   }
